requirements.txt
boto3==1.34.55
black==24.2.0
dpath==2.1.3
fabric==2.7.1
google_api_python_client==2.68.0
invoke==1.7.3
isort==5.13.2
keyring==23.13.1
keyrings.google-artifactregistry-auth==1.1.1
packaging==23.2
prompt_toolkit==3.0.31
pylint==2.17.2
pylint-exit==1.2.0
pytz==2022.1
PyYAML==6.0
requests==2.25.1
rich==12.6.0
protobuf==5.26.1
